rename pages directory and update imports
