+++ README.md
@@ -1,4 +1,10 @@
 # buildledger_2025
 
 [Edit in StackBlitz next generation editor ⚡️](https://stackblitz.com/~/github.com/juud-8/buildledger_2025)
-// trigger CI
+
+## Development
+
+1. Install dependencies with `npm install`.
+2. Start the development server with `npm run dev`.
+
+This project uses [Vite](https://vitejs.dev/) with React and TypeScript. Environment variables are read using `import.meta.env` (for example `import.meta.env.VITE_SUPABASE_URL`).
