+++ src/App.tsx
@@ -2,9 +2,9 @@ import React from 'react';
 import { BrowserRouter, Routes, Route, Navigate } from 'react-router-dom';
 import { AuthProvider } from './context/AuthContext';
 import ProtectedRoute from './components/ProtectedRoute';
-import Login from './pages/Login';
-import Register from './pages/Register';
-import Dashboard from './pages/Dashboard';
+import Login from './routes/Login';
+import Register from './routes/Register';
+import Dashboard from './routes/Dashboard';
 
 const App = () => (
   <BrowserRouter>
